Add selector support to json indexes
Adds selector support to json indexes. The selector can be used to filter what documents are
added to the index. When executing a query the index will
only be choosen if the query selector contains the selector defined in
the index or if the index is defined via `use_index`.