Fix json index selection
JSON index selection in Mango previously allowed an
index to be used if a range scan on the first component
of its compound key could be generated from the query selector.

For instance, if I had an index:

[A, B]

is_usable would return true for the query:

A > 5

This is incorrect because JSON indexes only index documents
that contain all the fields in the index; null values are ok,
but the field must be present. That means that for the above
query, the index would implicitly include only documents where
B exists, dropping documents which only included A from the
result set.

This commit changes is_usable so that it only returns true
if all the keys in the index are required to exist by the
selector. This means that in the worst case e.g. none of
the predicates can be used to generate a range query, we'd end up
performing a full index scan, but this is still more efficient
than a full database scan.

We leave the generation of the optimal range for a given index
as a separate exercise - currently this happens after index
selection.

Potentially we'd want to score indexes during index selection
based on their ability to restrict the result set, etc.