Covering indexes for Mango
Before querying the json index, determine whether the
index covers the fields requested in the query object.

If it does, do not pass include_docs=true in the view
parameters. Instead, construct a "fake" document using
the key of each row returned and pass this through the
rest of the query pipeline.