Update usage of couch_hash:md5 shim
https://github.com/apache/couchdb/pull/1445 introduced a shim to
enable CouchDB to be compiled to use the Erlang MD5 function.
This allows CouchDB to run in FIPS environments where the crypto
module is restricted such that `crypto:hash(md5,..)` is blocked
(fails with `notsup` error).

This commit replaces usage of `crypto:hash(md5, ..)` introduced since
the original PR with the shim function.